fix logical comparison with primitive types
when comparing different types, and the first operand is a primitive,
such as in `"hoge" == Cel::String("hoge")`, ruby will revert the order
and see if the second type implemennts "==", which Cel::String does. It
correctly returns `Bool.new(false)`.

However, ruby primitives interpret that as true, as that's not a ruby
"falsy", such as false or nil, so String.== converts it to true. Because
there is no `to_bool` coercion, we can't just implement it on bool, so
we have to fall back to returning the primitive bool when the operator
is a primmitive ruby type.